% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNewCCI.R
\name{FindNewCCI}
\alias{FindNewCCI}
\title{Identifying New Cell-Cell Interactions Using Latent Ligands}
\usage{
FindNewCCI(
  seu.obj,
  latent.ligands,
  source.celltype,
  target.celltype,
  celltype.colname = "Cell.Type",
  logFC.cutoff = 0.2,
  pct.dff.cutoff = 0.1,
  padj.cutoff = 0.05,
  CCI.database = TimeCCI::LRdf,
  species = "human"
)
}
\arguments{
\item{seu.obj}{A Seurat object containing the single-cell RNA-seq data.}

\item{latent.ligands}{A vector of latent ligand gene symbols that are hypothesized to be involved in cell-cell communication.}

\item{source.celltype}{The source cell type from which ligand signals are hypothesized to originate.}

\item{target.celltype}{The target cell type that receives signals via receptors.}

\item{celltype.colname}{The name of the column in \code{seu.obj@meta.data} that contains the cell type annotations. Default is 'Cell.Type'.}

\item{logFC.cutoff}{Numeric value specifying the log fold change threshold for filtering differentially expressed genes. Default is 0.2.}

\item{pct.dff.cutoff}{Numeric value for the percent difference cutoff between in-group and out-group cells. Default is 0.1.}

\item{padj.cutoff}{Numeric value specifying the adjusted p-value threshold for filtering significant interactions. Default is 0.05.}

\item{CCI.database}{The CellChat database to use for CCI analysis. Default is \code{TimeCCI::LRdf}.}

\item{species}{The species for the analysis. Default is "human".}
}
\value{
A CellChat object containing identified cell-cell interactions between the specified source and target cell types based on the given ligand-receptor pairs.
}
\description{
This function performs a detailed analysis of cell-cell interactions (CCI) in single-cell RNA sequencing data using latent ligand-receptor pairs. The function integrates cell type information and differentially expressed genes to identify potential cell-cell interactions. It uses the CellChat database and the wilcoxauc method from Presto for differential expression analysis.
}
\details{
This function operates as follows:
\enumerate{
\item Differential expression analysis is performed on the target cell type using the Wilcoxon rank-sum test via the \code{presto::wilcoxauc()} function.
\item Genes are filtered based on log fold change, percentage difference, and adjusted p-value cutoffs.
\item Receptor genes are identified from the latent ligand-receptor pairs and filtered from the differentially expressed genes.
\item A custom ligand-receptor interaction database is created using the provided latent ligands and the receptors identified from differentially expressed genes.
\item The function creates a \code{CellChat} object from the Seurat object’s expression data and metadata.
\item The custom ligand-receptor database is applied, and the CellChat pipeline is executed to identify overexpressed genes and interactions, compute communication probabilities, and aggregate interaction networks.
}

The function returns a CellChat object that can be further used for visualizing and interpreting cell-cell communication.
}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
